+++ src/main/java/com/uet/k62/web/system/examination/repository/RoleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface RoleRepository extends JpaRepository<Role, BigInteger> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
 }
